+++ src/main/java/com/idle/fmd/domain/lol/job/LolApiJob.java
@@ -6,6 +6,15 @@ import com.idle.fmd.domain.lol.dto.LolMatchDto;
 import com.idle.fmd.domain.lol.entity.LolAccountEntity;
 import com.idle.fmd.domain.lol.entity.LolInfoEntity;
 import com.idle.fmd.domain.lol.entity.LolMatchEntity;
+import com.idle.fmd.domain.lol.job.processor.LolAccountItemProcessor;
+import com.idle.fmd.domain.lol.job.processor.LolInfoItemProcessor;
+import com.idle.fmd.domain.lol.job.processor.LolMatchItemProcessor;
+import com.idle.fmd.domain.lol.job.reader.LolAccountItemReader;
+import com.idle.fmd.domain.lol.job.reader.LolInfoItemReader;
+import com.idle.fmd.domain.lol.job.reader.LolMatchItemReader;
+import com.idle.fmd.domain.lol.job.writer.LolAccountItemWriter;
+import com.idle.fmd.domain.lol.job.writer.LolInfoItemWriter;
+import com.idle.fmd.domain.lol.job.writer.LolMatchItemWriter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.*;
