+++ src/main/java/com/idle/fmd/domain/lol/job/reader/LolAccountItemReader.java
@@ -1,4 +1,4 @@
-package com.idle.fmd.domain.lol.job;
+package com.idle.fmd.domain.lol.job.reader;
 
 import com.idle.fmd.domain.lol.dto.LolAccountDto;
 import com.idle.fmd.domain.lol.entity.LolAccountEntity;
