+++ src/main/java/com/idle/fmd/domain/lol/job/writer/LolInfoItemWriter.java
@@ -1,4 +1,4 @@
-package com.idle.fmd.domain.lol.job;
+package com.idle.fmd.domain.lol.job.writer;
 
 import com.idle.fmd.domain.lol.entity.LolInfoEntity;
 import com.idle.fmd.domain.lol.repo.LolInfoRepository;
