+++ src/main/java/com/idle/fmd/domain/lol/job/writer/LolMatchItemWriter.java
@@ -1,4 +1,4 @@
-package com.idle.fmd.domain.lol.job;
+package com.idle.fmd.domain.lol.job.writer;
 
 import com.idle.fmd.domain.lol.entity.LolMatchEntity;
 import com.idle.fmd.domain.lol.repo.LolMatchRepository;
